feat(newgrf): honor sprite stack register 100

+++ crates/openttdrs-core/src/autoreplace.rs
@@ -568,7 +568,7 @@ mod tests {
             variable: 0x1A,
             param: None,
             adjust: Action2VarAdjust {
-                and_mask: value,
+                and_mask: u32::from(value),
                 ..Action2VarAdjust::default()
             },
         };
@@ -584,7 +584,7 @@ mod tests {
                     variable: 0x10,
                     param: None,
                     adjust: Action2VarAdjust {
-                        and_mask: u8::MAX,
+                        and_mask: u32::from(u8::MAX),
                         ..Action2VarAdjust::default()
                     },
                 },
